CRITICAL BUGFIX: fix base initialization
There was a critical error in json2sh which allows
code execution for certain workloads.

This was due to the freelist optimization (to skip allocs)
and improper re-initalization of already used structures.
This bug now is properly commented, so shall never occur again.

Please remember that this code here still is barely tested.